back-end/pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = "back end api service for the project"
authors = ["oliver-mio <ollie.mio@student.uts.edu.au>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.104.0"
debugpy = "^1.6.7"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
dependency-injector = "^4.41.0"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.20"
numpy = "^1.26.1"
sklearn = "^0.0.post10"
scikit-learn = "^1.3.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
hypothesis = "^6.82.3"
httpx = "^0.24.1"
pytest-coverage = "^0.0"


[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = [".", "service/src", "service/src/tests"]
filterwarnings = ["ignore::DeprecationWarning:pytest_asyncio.*:"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
